Added the arm joint controller
Had to tweak a large number of things. Discovered that I can't have multiple controller "launcher" nodes, it doesn't like it (or at least how I was trying to do it), so I made one overall "controller launcher" and "hardware interface" that does both the arm and drivebase. This isn't really an issue, it's just a bit less modularity.